Removed obsolete code
