+++ BthPS3/BusLogic.c
@@ -48,166 +48,9 @@ IoctlHandler_IoctlRecord G_PDO_IoctlSpecification[] =
 	{IOCTL_BTHPS3_HID_CONTROL_WRITE, 1, 0, BthPS3_PDO_HandleHidControlWrite},
 	{IOCTL_BTHPS3_HID_INTERRUPT_READ, 0, 1, BthPS3_PDO_HandleHidInterruptRead},
 	{IOCTL_BTHPS3_HID_INTERRUPT_WRITE, 1, 0, BthPS3_PDO_HandleHidInterruptWrite},
+	/* TODO: should we keep supporting this? It can be sent directly to the host radio instead... */
 	{IOCTL_BTH_DISCONNECT_DEVICE, sizeof(UINT64), 0, BthPS3_PDO_HandleBthDisconnect},
 };
-
-
-//
-// TODO: legacy, remove/migrate!
-// 
-
-//
-// Triggered on entering I/O idle state
-// 
-NTSTATUS BthPS3_PDO_EvtWdfDeviceD0Exit(
-	WDFDEVICE Device,
-	WDF_POWER_DEVICE_STATE TargetState
-)
-{
-	NTSTATUS                    status;
-	WDFDEVICE                   parentDevice;
-	WDFIOTARGET                 parentTarget;
-	PBTHPS3_PDO_DEVICE_CONTEXT  pPdoDevCtx;
-	WDFREQUEST					dcRequest;
-	WDF_OBJECT_ATTRIBUTES		attributes;
-	WDFMEMORY					payload;
-
-
-	UNREFERENCED_PARAMETER(Device);
-	UNREFERENCED_PARAMETER(TargetState);
-
-	FuncEntry(TRACE_BUSLOGIC);
-
-	pPdoDevCtx = GetPdoDeviceContext(Device);
-	parentDevice = WdfPdoGetParent(Device);
-	parentTarget = WdfDeviceGetIoTarget(parentDevice);
-
-	TraceInformation(
-		TRACE_BUSLOGIC,
-		"Requesting device disconnect"
-	);
-
-	/*
-	 * Low power conditions are not supported by the remote device,
-	 * if low power or idle was requested, instruct radio to disconnect.
-	 */
-
-	WDF_OBJECT_ATTRIBUTES_INIT(&attributes);
-	attributes.ParentObject = Device;
-
-	status = WdfRequestCreate(
-		&attributes,
-		parentTarget,
-		&dcRequest
-	);
-	if (!NT_SUCCESS(status))
-	{
-		TraceError(
-			TRACE_BUSLOGIC,
-			"WdfRequestCreate failed with status %!STATUS!",
-			status
-		);
-	}
-
-	WDF_OBJECT_ATTRIBUTES_INIT(&attributes);
-	attributes.ParentObject = dcRequest;
-
-	status = WdfMemoryCreatePreallocated(
-		&attributes,
-		&pPdoDevCtx->ClientConnection->RemoteAddress,
-		sizeof(BTH_ADDR),
-		&payload
-	);
-	if (!NT_SUCCESS(status))
-	{
-		TraceError(
-			TRACE_BUSLOGIC,
-			"WdfMemoryCreatePreallocated failed with status %!STATUS!",
-			status
-		);
-	}
-
-	status = WdfIoTargetFormatRequestForIoctl(
-		parentTarget,
-		dcRequest,
-		IOCTL_BTH_DISCONNECT_DEVICE,
-		payload,
-		NULL,
-		NULL,
-		NULL
-	);
-	if (!NT_SUCCESS(status))
-	{
-		TraceError(
-			TRACE_BUSLOGIC,
-			"WdfIoTargetFormatRequestForIoctl failed with status %!STATUS!",
-			status
-		);
-	}
-
-	WdfRequestSetCompletionRoutine(
-		dcRequest,
-		BthPS3_PDO_DisconnectRequestCompleted,
-		NULL
-	);
-
-	//
-	// Request parent to abandon us :'(
-	// 
-	if (WdfRequestSend(
-		dcRequest,
-		parentTarget,
-		WDF_NO_SEND_OPTIONS
-	) == FALSE)
-	{
-		status = WdfRequestGetStatus(dcRequest);
-
-		if (status != STATUS_DEVICE_NOT_CONNECTED && !NT_SUCCESS(status))
-		{
-			TraceError(
-				TRACE_BUSLOGIC,
-				"WdfRequestSend failed with status %!STATUS!",
-				status
-			);
-		}
-		else
-		{
-			//
-			// Override
-			// 
-			status = STATUS_SUCCESS;
-		}
-	}
-
-	FuncExit(TRACE_BUSLOGIC, "status=%!STATUS!", status);
-
-	return status;
-}
-
-//
-// Disconnect request got completed
-// 
-void BthPS3_PDO_DisconnectRequestCompleted(
-	_In_ WDFREQUEST Request,
-	_In_ WDFIOTARGET Target,
-	_In_ PWDF_REQUEST_COMPLETION_PARAMS Params,
-	_In_ WDFCONTEXT Context
-)
-{
-	UNREFERENCED_PARAMETER(Target);
-	UNREFERENCED_PARAMETER(Context);
-
-	FuncEntryArguments(
-		TRACE_BUSLOGIC,
-		"status=%!STATUS!",
-		Params->IoStatus.Status
-	);
-
-	WdfObjectDelete(Request);
-
-	FuncExitNoReturn(TRACE_BUSLOGIC);
-}
-
 
 
 //
